requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.7
djangorestframework==3.13.1
importlib-metadata==4.12.0
isort==5.10.1
Markdown==3.4.1
numpy==1.23.1
pandas==1.4.3
python-dateutil==2.8.2
python-markdown-math==0.8
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
zipp==3.8.1
